Automatically download API token if one is not stored
Generating the token no longer needs to be done separately. The script will detect if the API token is missing and try to request it.

The user needs to manually enter the username and password, which may be a problem for automatic scripts. (no worries, a fix is on the way)

If no username or password is given, the script exits early.

To reset the token, use the `--reset-token` key which will function just as if the user did not have a token specified.

Also (though this was added in the previous commit) use the `--token` key to specify a token manually, which is useful if you want to temporarily use a different GitHub account for an upload.